Administrative setting.

+++ install.sh
@@ -3,6 +3,8 @@
 sudo pip3 install -r requirements.txt
 sudo python3 setup.py install
 # Deleting trash files
-sudo rm -r build/
-sudo rm -r dist/
-sudo rm -r chatbotQuery.egg-info/
+if  [ -d "$build"]; then
+    sudo rm -r build/
+    sudo rm -r dist/
+    sudo rm -r chatbotQuery.egg-info/
+fi
